+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Removed all JSON specifics, including default parsing functionality
 - Adjusted `Endpoint::body` to return an `Option` on success
 - Require `std::error::Error` instead of `std::fmt::Display` for
   `Endpoint::ApiError` type